Revert "Add sharding support for Mongo store tests. (#792)"
This reverts commit bf8994c202a3aaa4363b8d61ca964b0926df256a.
