+++ versioned/tiered/mongodb/src/test/java/org/projectnessie/versioned/mongodb/TestMongoDBStore.java
@@ -15,17 +15,89 @@
  */
 package org.projectnessie.versioned.mongodb;
 
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.projectnessie.versioned.impl.AbstractTestStore;
 
 /**
- * A test class that contains MongoDB specific tests for unsharded MongoDB.
+ * A test class that contains MongoDB specific tests.
  */
-@ExtendWith(LocalMongoD.class)
+@ExtendWith(LocalMongo.class)
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-class TestMongoDBStore extends TestMongoDBStoreBase {
+class TestMongoDBStore extends AbstractTestStore<MongoDBStore> {
+  private static final String testDatabaseName = "mydb";
+  private String connectionString;
+
+  @BeforeAll
+  void init(String connectionString) {
+    this.connectionString = connectionString;
+  }
+
+  @AfterAll
+  void close() {
+    if (null != store) {
+      store.close();
+    }
+  }
+
+  /**
+   * Creates an instance of MongoDBStore on which tests are executed.
+   * @return the store to test.
+   */
+  @Override
+  protected MongoDBStore createStore() {
+    return new MongoDBStore(createConfig());
+  }
+
+  @Override
+  protected MongoDBStore createRawStore() {
+    return new MongoDBStore(createConfig());
+  }
+
   @Override
   protected long getRandomSeed() {
     return 8612341233543L;
+  }
+
+  @Override
+  protected void resetStoreState() {
+    store.resetCollections();
+  }
+
+  @Override
+  protected int loadSize() {
+    return MongoDBStore.LOAD_SIZE;
+  }
+
+  private MongoStoreConfig createConfig() {
+    return new MongoStoreConfig() {
+      @Override
+      public String getConnectionString() {
+        return connectionString;
+      }
+
+      @Override
+      public String getDatabaseName() {
+        return testDatabaseName;
+      }
+    };
+  }
+
+  // Disabled tests
+  @Override
+  protected boolean supportsDelete() {
+    return false;
+  }
+
+  @Override
+  protected boolean supportsUpdate() {
+    return false;
+  }
+
+  @Override
+  protected boolean supportsConditionExpression() {
+    return false;
   }
 }
