--- versioned/tiered/mongodb/src/test/java/org/projectnessie/versioned/mongodb/TestShardedMongoDBStore.java
@@ -1,31 +1,0 @@
-/*
- * Copyright (C) 2020 Dremio
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.projectnessie.versioned.mongodb;
-
-import org.junit.jupiter.api.TestInstance;
-import org.junit.jupiter.api.extension.ExtendWith;
-
-/**
- * A test class that contains MongoDB specific tests for sharded MongoDB.
- */
-@ExtendWith(LocalMongoS.class)
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-class TestShardedMongoDBStore extends TestMongoDBStoreBase {
-  @Override
-  protected long getRandomSeed() {
-    return 98324122436234L;
-  }
-}
